Removes S3 usage store to refer to ElasticSearch
This is in order to maintain caching of usage while referring to
ElasticSearch as a more up to date resource.
