+++ common-lib/src/main/scala/com/gu/mediaservice/lib/BaseStore.scala
@@ -50,5 +50,3 @@ abstract class BaseStore[TStoreKey, TStoreVal](bucket: String, credentials: AWSC
 
   def update(): Unit
 }
-
-
